assignment-3.1-RAG/requirements.txt
transformers
sentence-transformers
faiss-cpu
requests
beautifulsoup4
tqdm
pandas
scikit-learn
jupyter 